refactor(room-layout): migrate to immutable layout snapshot model

+++ backend/src/main/java/com/cinemaebooking/backend/showtime/application/dto/showtimeFormat/ShowtimeFormatResponse.java
@@ -3,10 +3,12 @@ package com.cinemaebooking.backend.showtime.application.dto.showtimeFormat;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class ShowtimeFormatResponse {
     private Long id;
     private String name;
-    private Long extraPrice;
+    private BigDecimal extraPrice;
 }
